agent/requirements.txt
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
pyjwt==2.8.0
websockets==12.0
cryptography==41.0.7
python-logging-loki==0.3.1
watchdog==3.0.0 